+++ cmsplugin_blocks/forms/slider.py
@@ -4,31 +4,40 @@ from django import forms
 from djangocms_text_ckeditor.widgets import TextEditorWidget
 
 from cmsplugin_blocks.models.slider import Slider, SlideItem
+from cmsplugin_blocks.widgets import ClearableFileInputButtonBase
 
 
 class SliderForm(forms.ModelForm):
     class Meta:
         model = Slider
         fields = [
-            'title',
-            'template',
+            "title",
+            "template",
         ]
         exclude = []
+
+    class Media:
+        css = {
+            "all": ("cmsplugin_blocks/css/admin/slider.css",),
+        }
+        js = ("cmsplugin_blocks/js/fileinputbutton.js",)
 
 
 class SlideItemForm(forms.ModelForm):
     class Meta:
         model = SlideItem
         widgets = {
-            'content': TextEditorWidget,
+            "image": ClearableFileInputButtonBase,
+            "content": TextEditorWidget,
         }
         fields = [
-            'slider',
-            'image',
-            'content',
-            'order',
-            'link_name',
-            'link_url',
-            'link_open_blank',
+            "slider",
+            "title",
+            "image",
+            "content",
+            "order",
+            "link_name",
+            "link_url",
+            "link_open_blank",
         ]
         exclude = []