+++ cmsplugin_blocks/utils/__init__.py
@@ -5,13 +5,13 @@ from cmsplugin_blocks.utils.validators import (
 )
 
 from cmsplugin_blocks.utils.smart_format import (
-    AVAILABLE_FORMATS, AVAILABLE_FORMAT_EXTENSIONS, SmartFormatMixin, SvgFile
+    AVAILABLE_FORMAT_EXTENSIONS, SmartFormatMixin, SvgFile
 )
 
 
 __all__ = [
     'store_images_from_zip',
     'is_valid_image_filename', 'validate_file_size', 'validate_zip',
-    'AVAILABLE_FORMATS', 'AVAILABLE_FORMAT_EXTENSIONS', 'SmartFormatMixin',
+    'AVAILABLE_FORMAT_EXTENSIONS', 'SmartFormatMixin',
     'SvgFile',
 ]